requirements.txt

pymongo~=4.0.2          # python client for mongodb

redis~=4.3.4
nest_asyncio
gunicorn~=20.1.0        #
pydantic~=1.9.0         #
loguru==0.6.0           # for logger
numpy~=1.26.4
watchdog~=2.1.7         # reload
python-dotenv~=0.19.2   # to store environment variables in .env file, we can store in system,
Celery~=5.2.6
flower==1.0.0
requests==2.27.1
pyOpenSSL==22.0.0
aioredis==2.0.1
pytest~=7.1.0
pytest-asyncio~=0.18.3
pytest-cov~=3.0.0
Twisted~=22.8.0
itemloaders~=1.0.6
w3lib~=2.0.1
cryptography~=38.0.3
six~=1.16.0
bkit~=0.0.8

sentence-transformers~=3.3.1
scikit-learn~=1.6.1
chromadb~=1.0.15
torch~=2.5.1
transformers~=4.47.1
asyncpg~=0.24.0